snap-recovery: rename volmgr package to partition
