+++ cmd/snap-recovery/partition/sfdisk.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package volmgr
+package partition
 
 import (
 	"bytes"
